Convert map builtin to class
