+++ Src/IronPython/Modules/Builtin.cs
@@ -802,43 +802,10 @@ Noteworthy: None is the `nil' object; Ellipsis represents `...' in slices.";
             );
         }
 
-        [PythonType("map")]
-        public class mapobject : IEnumerable {
-            private CodeContext _context;
-            private object _func;
-            private object[] _iters;
-
-            public mapobject(CodeContext context, object func, params object[] iters) {
-                _context = context;
-                _func = func;
-                _iters = iters;
+        public static PythonType map {
+            get {
+                return DynamicHelpers.GetPythonTypeFromType(typeof(Map));
             }
-
-            IEnumerator IEnumerable.GetEnumerator() {
-                IEnumerator[] enumerators = _iters.Select(x => PythonOps.GetEnumerator(x)).ToArray();
-                while (enumerators.All(x => x.MoveNext())) {
-                    yield return PythonOps.CallWithContext(_context, _func, enumerators.Select(x => x.Current).ToArray());
-                }
-            }            
-        }
-
-        public static object map(CodeContext/*!*/ context, object func, params object[] iters) {
-            if(iters.Length == 0) {
-                throw PythonOps.TypeError("map() must have at least two arguments.");
-            }
-
-            if(!PythonOps.IsCallable(context, func)) {
-                throw PythonOps.TypeError("'{0}' object is not callable", PythonOps.GetPythonTypeName(func));
-            }
-
-            foreach(object o in iters) {
-                IEnumerator e;
-                if(!PythonOps.TryGetEnumerator(context, o, out e)) {
-                    throw PythonOps.TypeError("'{0}' object is not iterable", PythonOps.GetPythonTypeName(o));
-                }
-            }
-
-            return new mapobject(context, func, iters);
         }
 
         private static object UndefinedKeywordArgument = new object();
